# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.BillOfMaterial:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      product_id:
        type: integer
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  entity.ManufacturingFacility:
    properties:
      address:
        type: string
      capacity:
        description: Production capacity per day
        type: integer
      created_at:
        type: string
      id:
        type: integer
      manager:
        type: string
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.ProductionOrder:
    properties:
      completed_qty:
        type: integer
      created_at:
        type: string
      deadline:
        type: string
      defect_qty:
        type: integer
      facility_id:
        type: integer
      id:
        type: integer
      notes:
        type: string
      product_id:
        type: integer
      quantity:
        type: integer
      start_date:
        type: string
      status:
        $ref: '#/definitions/entity.ProductionOrderStatus'
      updated_at:
        type: string
    type: object
  entity.ProductionOrderStatus:
    enum:
    - pending
    - in_process
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusInProcess
    - OrderStatusCompleted
    - OrderStatusCancelled
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/lugondev/erp-warehouse-simple
  description: Simple ERP Warehouse Management System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ERP Warehouse API
  version: "1.0"
paths:
...
  /api/v1/manufacturing/bom:
    post:
      consumes:
      - application/json
      description: Create bill of materials for a product
      parameters:
      - description: BOM details with items
        in: body
        name: bom
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.BillOfMaterial'
      summary: Create BOM
      tags:
      - Manufacturing
  /api/v1/manufacturing/facilities:
    get:
      description: Get list of all manufacturing facilities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ManufacturingFacility'
            type: array
      summary: List facilities
      tags:
      - Manufacturing
    post:
      consumes:
      - application/json
      description: Create new manufacturing facility
      parameters:
      - description: Facility details
        in: body
        name: facility
        required: true
        schema:
          $ref: '#/definitions/entity.ManufacturingFacility'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.ManufacturingFacility'
      summary: Create manufacturing facility
      tags:
      - Manufacturing
  /api/v1/manufacturing/facilities/{id}:
    get:
      description: Get manufacturing facility details by ID
      parameters:
      - description: Facility ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ManufacturingFacility'
      summary: Get facility by ID
      tags:
      - Manufacturing
  /api/v1/manufacturing/orders:
    post:
      consumes:
      - application/json
      description: Create new production order
      parameters:
      - description: Production order details
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/entity.ProductionOrder'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.ProductionOrder'
      summary: Create production order
      tags:
      - Manufacturing
  /api/v1/manufacturing/orders/{id}/progress:
    put:
      consumes:
      - application/json
      description: Update production progress for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Progress details
        in: body
        name: progress
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Progress updated
          schema:
            type: string
      summary: Update production progress
      tags:
      - Manufacturing
  /api/v1/manufacturing/orders/{id}/start:
    post:
      description: Start production for an order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Production started
          schema:
            type: string
      summary: Start production
      tags:
      - Manufacturing
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"